+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -4,13 +4,13 @@ import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import java.util.Optional;
 import java.util.UUID;
-import org.apache.controller.AbstractController;
+import org.apache.coyote.controller.AbstractController;
+import org.apache.coyote.http.Cookie;
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.response.HttpResponse;
 import org.apache.coyote.http11.Http11Processor;
-import org.apache.http.Cookie;
-import org.apache.http.request.HttpRequest;
-import org.apache.http.response.HttpResponse;
-import org.apache.session.Session;
-import org.apache.session.SessionManager;
+import org.apache.coyote.session.Session;
+import org.apache.coyote.session.SessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
