+++ tomcat/src/main/java/com/techcourse/controller/RootController.java
@@ -1,10 +1,10 @@
 package com.techcourse.controller;
 
-import org.apache.controller.AbstractController;
-import org.apache.http.request.HttpRequest;
-import org.apache.http.response.HttpResponse;
-import org.apache.http.value.HttpHeader;
-import org.apache.http.value.StatusCode;
+import org.apache.coyote.controller.AbstractController;
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.response.HttpResponse;
+import org.apache.coyote.http.value.HttpHeader;
+import org.apache.coyote.http.value.StatusCode;
 
 public class RootController extends AbstractController {
