--- tomcat/src/main/java/org/apache/controller/Controller.java
@@ -1,11 +1,0 @@
-package org.apache.controller;
-
-import org.apache.http.request.HttpRequest;
-import org.apache.http.response.HttpResponse;
-
-public interface Controller {
-
-    boolean canProcessable(HttpRequest request);
-
-    void service(HttpRequest request, HttpResponse response) throws Exception;
-}
