+++ tomcat/src/main/java/org/apache/coyote/controller/Controller.java
@@ -1,0 +1,11 @@
+package org.apache.coyote.controller;
+
+import org.apache.coyote.http.request.HttpRequest;
+import org.apache.coyote.http.response.HttpResponse;
+
+public interface Controller {
+
+    boolean canProcessable(HttpRequest request);
+
+    void service(HttpRequest request, HttpResponse response) throws Exception;
+}
