+++ tomcat/src/main/java/org/apache/coyote/exception/DataNotFoundException.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.exception;
+
+public class DataNotFoundException extends RuntimeException {
+
+    public DataNotFoundException(String message) {
+        super(message);
+    }
+}
