+++ tomcat/src/main/java/org/apache/coyote/exception/FileReadException.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.exception;
+
+public class FileReadException extends RuntimeException {
+
+    public FileReadException(String message) {
+        super(message);
+    }
+}
