+++ tomcat/src/main/java/org/apache/coyote/exception/InvalidRequestException.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.exception;
+
+public class InvalidRequestException extends RuntimeException {
+
+    public InvalidRequestException(String message) {
+        super(message);
+    }
+}
