+++ tomcat/src/main/java/org/apache/coyote/exception/RequestProcessingException.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.exception;
+
+public class RequestProcessingException extends RuntimeException {
+
+    public RequestProcessingException(String message) {
+        super(message);
+    }
+}
