+++ tomcat/src/main/java/org/apache/coyote/exception/SocketReadException.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.exception;
+
+public class SocketReadException extends RuntimeException {
+
+    public SocketReadException(String message) {
+        super(message);
+    }
+}
