+++ tomcat/src/main/java/org/apache/coyote/exception/SocketWriteException.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.exception;
+
+public class SocketWriteException extends RuntimeException {
+
+    public SocketWriteException(String message) {
+        super(message);
+    }
+}
