+++ tomcat/src/main/java/org/apache/coyote/http/value/HttpMethod.java
@@ -1,0 +1,5 @@
+package org.apache.coyote.http.value;
+
+public enum HttpMethod {
+    GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS, CONNECT
+}
