+++ tomcat/src/main/java/org/apache/coyote/session/Session.java
@@ -1,0 +1,22 @@
+package org.apache.coyote.session;
+
+import com.techcourse.model.User;
+
+public class Session {
+
+    private final String id;
+    private final User user;
+
+    public Session(String id, User user) {
+        this.id = id;
+        this.user = user;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public User getUser() {
+        return user;
+    }
+}
