--- tomcat/src/main/java/org/apache/exception/DataNotFoundException.java
@@ -1,8 +1,0 @@
-package org.apache.exception;
-
-public class DataNotFoundException extends RuntimeException {
-
-    public DataNotFoundException(String message) {
-        super(message);
-    }
-}
