--- tomcat/src/main/java/org/apache/exception/FileReadException.java
@@ -1,8 +1,0 @@
-package org.apache.exception;
-
-public class FileReadException extends RuntimeException {
-
-    public FileReadException(String message) {
-        super(message);
-    }
-}
