--- tomcat/src/main/java/org/apache/exception/InvalidRequestException.java
@@ -1,8 +1,0 @@
-package org.apache.exception;
-
-public class InvalidRequestException extends RuntimeException {
-
-    public InvalidRequestException(String message) {
-        super(message);
-    }
-}
