--- tomcat/src/main/java/org/apache/exception/RequestProcessingException.java
@@ -1,8 +1,0 @@
-package org.apache.exception;
-
-public class RequestProcessingException extends RuntimeException {
-
-    public RequestProcessingException(String message) {
-        super(message);
-    }
-}
