--- tomcat/src/main/java/org/apache/exception/SocketReadException.java
@@ -1,8 +1,0 @@
-package org.apache.exception;
-
-public class SocketReadException extends RuntimeException {
-
-    public SocketReadException(String message) {
-        super(message);
-    }
-}
