--- tomcat/src/main/java/org/apache/exception/SocketWriteException.java
@@ -1,8 +1,0 @@
-package org.apache.exception;
-
-public class SocketWriteException extends RuntimeException {
-
-    public SocketWriteException(String message) {
-        super(message);
-    }
-}
