--- tomcat/src/main/java/org/apache/http/value/HttpMethod.java
@@ -1,5 +1,0 @@
-package org.apache.http.value;
-
-public enum HttpMethod {
-    GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS, CONNECT
-}
