--- tomcat/src/main/java/org/apache/session/SessionManager.java
@@ -1,25 +1,0 @@
-package org.apache.session;
-
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-public class SessionManager {
-
-    private static final Map<String, Session> SESSIONS = new ConcurrentHashMap<>();
-    //TODO: 세션 만료 및 정리 정책 고려 (2025-09-9, 화, 14:1)
-
-    public static void add(Session session) {
-        SESSIONS.put(session.getId(), session);
-    }
-
-    public static Session findSession(String sessionId) {
-        return SESSIONS.get(sessionId);
-    }
-
-    public static void remove(final String id) {
-        SESSIONS.remove(id);
-    }
-
-    private SessionManager() {
-    }
-}
